torch/config: fix mock behaviour
Mock only replaces the value that was removed, if after deletion, it
does not see the attribute.

The core code which is tripping over this bug is https://github.com/python/cpython/blob/main/Lib/unittest/mock.py#L1637
which is working around the proxy object django settings, to avoid
leavining a trailing past value.

Some discussion is at https://bugs.python.org/issue22138

I expect it's possible to fix this in python by going (if getattr(foo)
!= original, setattr(foo, original)). I am going to talk to python
about this, but for now just fix this.

ghstack-source-id: 15f5860be12ebb2c5f1305f67ea7de3395f2e365
Pull Request resolved: https://github.com/pytorch/pytorch/pull/140779